app/requirements.txt
streamlit
tensorflow
numpy
Pillow
fpdf
googletrans==4.0.0-rc1
requests